Fix error reporting
